backend/bio-sync-app/requirements.txt
PyQt6>=6.6.0
mysql-connector-python>=8.3.0
pyzk>=0.9
pycryptodome>=3.20.0
requests>=2.31.0
Pillow>=10.0.0
pyinstaller>=6.6.0
